boolean hasMoved, KingStrategy, Castle
boolean hasMoved has been implemented in order for pieces to have an initialPosition. By doing this we can now if the piece has moved or not.
KingStrategy no longer exist, the King now take use of all the following strategies:
Vertical, Horizontal and Diagonal.

Castle logic is starting to come to life.